implemented exception handling

+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/CredentialController.java
@@ -1,11 +1,11 @@
 package com.udacity.jwdnd.course1.cloudstorage.controller;
 
+import com.udacity.jwdnd.course1.cloudstorage.exception.InvalidUserException;
 import com.udacity.jwdnd.course1.cloudstorage.model.Credential;
 import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import com.udacity.jwdnd.course1.cloudstorage.service.CredentialService;
 import com.udacity.jwdnd.course1.cloudstorage.service.EncryptionService;
 import com.udacity.jwdnd.course1.cloudstorage.service.UserService;
-import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -43,24 +43,33 @@ public class CredentialController {
 
         credentialService.saveCredential(credential);
         model.addAttribute("success", true);
+
         return "result";
     }
 
     @PostMapping("/credentials/delete")
-    public String deleteCredential(@ModelAttribute("credential") Credential credential, Model model) {
+    public String deleteCredential(@ModelAttribute("credential") Credential credential, Model model, Authentication auth) throws InvalidUserException {
+        User currentUser = userService.getUser(auth.getName());
+        Credential credentialToDelete = credentialService.getCredentialById(credential.getId());
+
+        if (!credentialToDelete.getUserId().equals(currentUser.getId())) {
+            throw new InvalidUserException("Credential does not belong to current user.");
+        }
+
         credentialService.deleteCredential(credential.getId());
         model.addAttribute("success", true);
+
         return "result";
     }
 
     @GetMapping("/credentials/decrypt/{credentialId}")
     @ResponseBody
-    public String getDecryptedPassword(@PathVariable int credentialId, Authentication auth) throws PermissionDeniedDataAccessException {
+    public String getDecryptedPassword(@PathVariable int credentialId, Authentication auth) throws InvalidUserException {
         User currentUser = userService.getUser(auth.getName());
         Credential credential = credentialService.getCredentialById(credentialId);
 
         if (!credential.getUserId().equals(currentUser.getId())) {
-            throw new PermissionDeniedDataAccessException("Credential does not belong to current user", null);
+            throw new InvalidUserException("Credential does not belong to current user.");
         }
 
         return encryptionService.decryptValue(credential.getPassword(), credential.getKey());
